Update smart_imaging.bat
Added a check if the directory for the SMART report exists. If it doesn't, the script will allow you to try again so you can create one. 

Added a check if the XWF "[filepath]\[filename].[format]" exists. If it exists, the script will call you a silly goose, but will let you try again and can change it.  Silly goose.